// angular20/src/app/swagger/model/okStatus.ts
/**
 * TourPlanner
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OkStatus { 
    isOk?: boolean;
    val?: string | null;
    error?: string | null;
}