// angular20/src/app/swagger/model/tourLogRequest.ts
/**
 * TourPlanner
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TourLogRequest { 
    logDateTime?: string;
    comment?: string | null;
    difficulty: string;
    totalDistanceKm?: number;
    totalTimeMinutes?: number;
    rating?: number | null;
}