// angular20/src/app/swagger/model/tourLogResponse.ts
/**
 * TourPlanner
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TourLogResponse { 
    id?: number;
    tourId?: number;
    logDateTime?: string;
    comment?: string | null;
    difficulty?: string | null;
    totalDistanceKm?: number;
    totalTimeMinutes?: number;
    rating?: number | null;
    createdAt?: string;
}